% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnsemble-package.R, R/SAForest-package.R
\docType{package}
\name{SAEnsemble-package}
\alias{SAEnsemble}
\alias{SAEnsemble-package}
\title{SAEnsemble: Simulated Annealing Forest}
\description{
prediction by SAEnsemble.

prediction by SAEnsemble.
}
\author{
\strong{Maintainer}: Minoru Inamori \email{inamori@ut-biomet.org}


\strong{Maintainer}: Minoru Inamori \email{inamori@ut-biomet.org}

}
